+++ bittytax/conv/parsers/coinbase.py
@@ -5,6 +5,7 @@ import sys
 import re
 from decimal import Decimal
 
+from ...config import config
 from ..out_record import TransactionOutRecord
 from ..dataparser import DataParser
 from ..exceptions import UnexpectedTypeError, UnexpectedContentError
@@ -18,14 +19,26 @@ def parse_coinbase(data_row, parser, _filename):
 
     if in_row[1] == "Receive":
         if "Coinbase Referral" in in_row[8]:
-            t_type = TransactionOutRecord.TYPE_GIFT_RECEIVED
+            # We can calculate the exact buy_value from the spot price
+            data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_GIFT_RECEIVED,
+                                                     data_row.timestamp,
+                                                     buy_quantity=in_row[3],
+                                                     buy_asset=in_row[2],
+                                                     buy_value=Decimal(in_row[4]) * \
+                                                               Decimal(in_row[3]),
+                                                     wallet=WALLET)
         else:
-            t_type = TransactionOutRecord.TYPE_DEPOSIT
-
-        data_row.t_record = TransactionOutRecord(t_type,
+            data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_DEPOSIT,
+                                                     data_row.timestamp,
+                                                     buy_quantity=in_row[3],
+                                                     buy_asset=in_row[2],
+                                                     wallet=WALLET)
+    elif in_row[1] in ("Coinbase Earn", "Rewards Income"):
+        data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_INCOME,
                                                  data_row.timestamp,
                                                  buy_quantity=in_row[3],
                                                  buy_asset=in_row[2],
+                                                 buy_value=in_row[6],
                                                  wallet=WALLET)
     elif in_row[1] == "Send":
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_WITHDRAWAL,
@@ -38,12 +51,14 @@ def parse_coinbase(data_row, parser, _filename):
         if currency is None:
             raise UnexpectedContentError(8, parser.in_header[8], in_row[8])
 
-        if Decimal(in_row[7]) == 0:
-            # No fee indicates Buy was for an early Referral reward
+        if config.coinbase_zero_fees_are_gifts and Decimal(in_row[7]) == 0:
+            # Zero fees "may" indicate an early referral reward, or airdrop
             data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_GIFT_RECEIVED,
                                                      data_row.timestamp,
                                                      buy_quantity=in_row[3],
                                                      buy_asset=in_row[2],
+                                                     buy_value=in_row[6] if Decimal(in_row[6]) > 0
+                                                                         else None,
                                                      wallet=WALLET)
         else:
             data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_TRADE,
@@ -69,8 +84,35 @@ def parse_coinbase(data_row, parser, _filename):
                                                  fee_quantity=in_row[7],
                                                  fee_asset=currency,
                                                  wallet=WALLET)
+    elif in_row[1] == "Convert":
+        convert_info = get_convert_info(in_row[8])
+        if convert_info is None:
+            raise UnexpectedContentError(8, parser.in_header[8], in_row[8])
+
+        buy_quantity = convert_info[2]
+        buy_asset = convert_info[3]
+        data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_TRADE,
+                                                 data_row.timestamp,
+                                                 buy_quantity=buy_quantity,
+                                                 buy_asset=buy_asset,
+                                                 buy_value=in_row[6],
+                                                 sell_quantity=in_row[3],
+                                                 sell_asset=in_row[2],
+                                                 sell_value=in_row[6],
+                                                 wallet=WALLET)
+
     else:
         raise UnexpectedTypeError(1, parser.in_header[1], in_row[1])
+
+def get_convert_info(notes):
+    if sys.version_info[0] < 3:
+        notes = notes.decode('utf8')
+
+    match = re.match(r'^Converted ([\d|,]*\.\d+) (\w+) to ([\d|,]*\.\d+) (\w+) *$', notes)
+
+    if match:
+        return match.groups()
+    return None
 
 def get_currency(notes):
     if sys.version_info[0] < 3: