+++ CHANGELOG.md
@@ -1,14 +1,45 @@
 # Change Log
 ## [Unreleased]
 ### Fixed
+- Accounting tool: "xlrd.biffh.XLRDError: Excel xlsx file; not supported" Exception. ([#36](https://github.com/BittyTax/BittyTax/issues/36))
+- Coinbase parser: added support for Convert transactions ([#46](https://github.com/BittyTax/BittyTax/issues/46))
+- Coinbase parser: mis-classifying trade as gift-received ([#47](https://github.com/BittyTax/BittyTax/issues/47))
+### Added
+- Etherscan parser: added internal transactions export.
+- Binance parser: added cash deposit and withdrawal exports.
+- Binance parser: added statements export.
+- Bitfinex parser: new "Trades" data file format added. ([#41](https://github.com/BittyTax/BittyTax/issues/41))
+- Bittrex parser: new deposits data file format added.
+- Coinbase parser: new config "coinbase_zero_fees_are_gifts" added.
+- Accounting/Conversion tool: support for milli/microsecond timestamps.
+### Changed
+- Conversion tool: UnknownAddressError exception changed to generic DataFilenameError.
+- Binance parser: use filename to determine if deposits or withdrawals.
+- Binance parser: updated quote assets via new script.
+- Crypto.com parser: added new "Supercharger" transaction types. ([#38](https://github.com/BittyTax/BittyTax/issues/38))
+- Coinbase parser: added Coinbase Earn/Rewards Income transactions.
+- Coinbase parser: get value (from spot price) where possible.
+- Bittrex parser: added market buy/sell transactions.
+- Ledger Live parser: fees now optional, as missing from ERC-20 wallets.
+- Bitstamp parser: fees now optional.
+- Accounting tool: same day pooling debug now only shows the pooled transactions.
+
+## Version [0.4.3] Beta (2020-12-04)
+Important:- if upgrading, please remove your price data cache file for CryptoCompare: `~/.bittytax/cache/CryptoCompare.json` (see Issue [#29](https://github.com/BittyTax/BittyTax/issues/29))
+### Fixed
 - UserWarning: Must have at least one data row in in add_table().
 - AttributeError: 'module' object has no attribute 'UTC'. ([#27](https://github.com/BittyTax/BittyTax/issues/27))
+- Crypto.com parser: fix date parser.
+- Incorrect price data for stablecoins via CryptoCompare. ([#29](https://github.com/BittyTax/BittyTax/issues/29))
 ### Added
 - Conversion tool: added parser for CGTCalculator.
-- Accounting/Conversion tool: support for milli/microsecond timestamps.
+- Conversion tool: added parser for Nexo.
+- Conversion tool: added parser for Kraken.
+- HitBTC parser: new data file format added.
 ### Changed
 - Hotbit parser: Negative fees are now set to zero.
 - Accounting tool: Drop buy/sell/fee transactions of zero quantity.
+- Crypto.com parser: Add support for referral_gift transaction type.
 
 ## Version [0.4.2] Beta (2020-10-30)
 ### Fixed
@@ -174,11 +205,11 @@
 - Conversion tool raises warning if 15-digit precision exceeded (Excel limit).
 - Conversion tool: added option to output in Recap import CSV format.
 ### Removed
-- Negative balance warning in a Section 104 holding. 
+- Negative balance warning in a Section 104 holding.
 - Logging removed from within config module.
 ### Changed
 - Logging is now initialised by each tool, instead of within the `config.py` module.
-- Conversion tool now outputs logging to `stderr` so it will be filtered when piping into `bittytax`. 
+- Conversion tool now outputs logging to `stderr` so it will be filtered when piping into `bittytax`.
 - The `pricedata.py` module has been renamed `valueasset.py`, and main function moved to new `price.py` module.
 - Package layout restructured, added subfolders for price and conv tools.
 - Refactored code for "all_handler" data parsers.
@@ -205,7 +236,8 @@ This is the initial beta release. Although it has been throughly tested, it's po
 - Exchange data files supported: Bitstamp, Bittrex, ChangeTip, Circle, Coinbase, Coinbase Pro, Coinfloor, Cryptopia, Cryptsy, Gatehub, OKEx, Poloniex, TradeSatoshi, Uphold.
 - Explorer data files supported: Etherscan.
 
-[Unreleased]: https://github.com/BittyTax/BittyTax/compare/v0.4.2...HEAD
+[Unreleased]: https://github.com/BittyTax/BittyTax/compare/v0.4.3...HEAD
+[0.4.3]: https://github.com/BittyTax/BittyTax/compare/v0.4.2...v0.4.3
 [0.4.2]: https://github.com/BittyTax/BittyTax/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/BittyTax/BittyTax/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/BittyTax/BittyTax/compare/v0.3.3...v0.4.0